feat: exceptions

+++ Tickets-API.Application/Handlers/DeleteTicketHandler.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Tickets_API.Application.Commands;
+using Tickets_API.Application.Exceptions;
 using Tickets_API.Domain.Interfaces;
 
 namespace Tickets_API.Application.Handlers
@@ -20,6 +21,12 @@ namespace Tickets_API.Application.Handlers
 
         public async Task<Unit> Handle(DeleteTicketCommand request, CancellationToken cancellationToken)
         {
+            var ticket = await _repository.GetByIdAsync(request.Id);
+            if (ticket == null)
+            {
+                throw new NotFoundException($"Chamado não encontrado.");
+            }
+               
             await _repository.DeleteAsync(request.Id);
             return Unit.Value;
         }
